[refactor]: replace custom privacy URL validation with shared utility function
