+++ frontend/src/shared/lib/index.ts
@@ -1,2 +1,3 @@
 export * from './bridge-adaptivity'
 export * from './date'
+export * from './privacy-url-validator'
